Fix issues with `constructed_cluster_name` length
Issues were presenting themselves when both the `cluster_name` and
`cluster_name_version` were _both_ 19 or 20 characters in length.

If they were both 19 characters in length, `substring_length` would
evaluate to `0`, meaning that the `substr(var.cluster_name, 0,
local.substring_length))` expression for `constructed_cluster_name`
would evaluate to `""` and end up making the `"-"` the first character
of the `constructed_cluster_name`, which is invalid.

If they were both 20 characters in length, the `substring_length` value
would evaluate to `-1`, which makes all `substr(...)` expressions using
it as the length parameter actually extract _all_ of the string.

This solution will:
    1. Use a `replace(...)` call on the first component of
       the `join("-", list(...))` expression to remove the leading `-` if
       present.
    2. Limit the overall length of the `constructed_cluster_name`,
       post-construction, to a maximum of 20 characters.
    3. Use `truncated_constructed_cluster_name` to avoid having to apply
       the truncation on every usage.